.av output added.
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/wopr@2758 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ src/runrunrun.h
@@ -72,7 +72,7 @@ int server( Logfile&, Config& );
 int test( Logfile&, Config& );
 int smooth_old( Logfile&, Config& );
 int smooth( Logfile&, Config& );
-unsigned long anahash( std::string& );
+unsigned long long anahash( std::string& );
 double word_entropy( std::map<std::string,int>& );
 int read_a3( Logfile&, Config& );
 int pplx( Logfile&, Config& );
